backend/relay/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
google-cloud-storage==2.19.0
google-genai>=1.56.0
pydantic==2.10.4
python-dotenv==1.0.1
